Fix ground initial position
Physics on client side got messed up because ground was not initially
where it was supposed to be. This is due to the fact that placement
of objects in p2 should be relative to their center (0.5,0.5) and not
to the top-left corner like you normally would do in 2D games.